backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
supabase==2.6.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
httpx==0.27.2
requests==2.32.3
groq==0.33.0
langchain_core==1.0.2
langchain_chroma==1.0.0
langchain_ollama==1.0.0
langchain_text_splitters==1.0.0

